Move mixin to utility module
